projects/accounting-automation/src/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.95.0
anthropic>=0.8.0
openai>=1.0.0
pillow>=10.0.0
requests>=2.31.0
